requirements.txt
Django==4.2.7
psycopg2-binary==2.9.9
Pillow==10.0.1
openpyxl==3.1.2
django-crispy-forms==2.0
crispy-bootstrap5==0.7
python-decouple==3.8
django-extensions==3.2.3
django-debug-toolbar==4.2.0
whitenoise==6.6.0
gunicorn==21.2.0
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
django-filter==23.3
reportlab==4.0.4
python-dateutil==2.8.2
bleach==6.1.0
django-colorfield==0.10.1
django-tinymce==3.6.1
django-import-export==3.3.1
django-activity-stream==2.0.0
django-notifications-hq==1.8.3